Make php-fpm forward environment variables
By default php-fpm has `clear_env` option activated.

This means that environment variables never will be passed on from the
system environment to the actual worker process for a given request.

This also means that Grocy will not be configureable via Environment
variables. Which is very common behaviour.

We can fix this by disabling the `clear_env`.

You can test that it works by for example changing the currency via an
environment variable:

  docker run -e GROCY_CURRENCY=NOK -p 8080:80 linuxserver/grocy:latest

Then opening:

  http://localhost:8080/stockoverview

And verifying that the app allows you to change the setting. There's
more info on how Grocy handles configuration, here:

  https://github.com/grocy/grocy/blob/master/config-dist.php#L3